+++ lib/buildingsync/model_articulation/spatial_element.rb
@@ -37,12 +37,13 @@
 require 'openstudio'
 require 'fileutils'
 require 'json'
-require_relative '../tmp_os_lib_model_generation_bricr'
+# require_relative '../tmp_os_lib_model_generation_bricr'
+require 'openstudio/extension/core/os_lib_model_generation'
 
 module BuildingSync
   # base class for objects that will configure workflows based on building sync files
   class SpatialElement
-    include Tmp_OsLib_ModelGenerationBRICR
+    #include Tmp_OsLib_ModelGenerationBRICR
 
     def initialize
       @total_floor_area = nil
@@ -134,6 +135,268 @@ module BuildingSync
         return occupancy_type
       end
     end
+
+    # create hash of space types and generic ratios of building floor area
+    def get_space_types_from_building_type(building_type, template, whole_building = true)
+
+      hash = {}
+
+      # TODO: - Confirm that these work for all standards
+
+      if building_type == 'SecondarySchool'
+        hash['Auditorium'] = { ratio: 0.0504, space_type_gen: true, default: false }
+        hash['Cafeteria'] = { ratio: 0.0319, space_type_gen: true, default: false }
+        hash['Classroom'] = { ratio: 0.3528, space_type_gen: true, default: true }
+        hash['Corridor'] = { ratio: 0.2144, space_type_gen: true, default: false }
+        hash['Gym'] = { ratio: 0.1646, space_type_gen: true, default: false }
+        hash['Kitchen'] = { ratio: 0.0110, space_type_gen: true, default: false }
+        hash['Library'] = { ratio: 0.0429, space_type_gen: true, default: false } # not in prototype
+        hash['Lobby'] = { ratio: 0.0214, space_type_gen: true, default: false }
+        hash['Mechanical'] = { ratio: 0.0349, space_type_gen: true, default: false }
+        hash['Office'] = { ratio: 0.0543, space_type_gen: true, default: false }
+        hash['Restroom'] = { ratio: 0.0214, space_type_gen: true, default: false }
+      elsif building_type == 'PrimarySchool'
+        hash['Cafeteria'] = { ratio: 0.0458, space_type_gen: true, default: false }
+        hash['Classroom'] = { ratio: 0.5610, space_type_gen: true, default: true }
+        hash['Corridor'] = { ratio: 0.1633, space_type_gen: true, default: false }
+        hash['Gym'] = { ratio: 0.0520, space_type_gen: true, default: false }
+        hash['Kitchen'] = { ratio: 0.0244, space_type_gen: true, default: false }
+        # TODO: - confirm if Library is 0.0 for all templates
+        hash['Library'] = { ratio: 0.0, space_type_gen: true, default: false }
+        hash['Lobby'] = { ratio: 0.0249, space_type_gen: true, default: false }
+        hash['Mechanical'] = { ratio: 0.0367, space_type_gen: true, default: false }
+        hash['Office'] = { ratio: 0.0642, space_type_gen: true, default: false }
+        hash['Restroom'] = { ratio: 0.0277, space_type_gen: true, default: false }
+      elsif building_type == 'SmallOffice'
+        # TODO: - populate Small, Medium, and Large office for whole_building false
+        if whole_building
+          hash['WholeBuilding - Sm Office'] = { ratio: 1.0, space_type_gen: true, default: true }
+        else
+          hash['SmallOffice - Breakroom'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['SmallOffice - ClosedOffice'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['SmallOffice - Conference'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['SmallOffice - Corridor'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['SmallOffice - Elec/MechRoom'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['SmallOffice - IT_Room'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['SmallOffice - Lobby'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['SmallOffice - OpenOffice'] = { ratio: 0.99, space_type_gen: true, default: true }
+          hash['SmallOffice - Restroom'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['SmallOffice - Stair'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['SmallOffice - Storage'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['SmallOffice - Classroom'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['SmallOffice - Dining'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['WholeBuilding - Sm Office'] = { ratio: 0.0, space_type_gen: true, default: false }
+        end
+      elsif building_type == 'MediumOffice'
+        if whole_building
+          hash['WholeBuilding - Md Office'] = { ratio: 1.0, space_type_gen: true, default: true }
+        else
+          hash['MediumOffice - Breakroom'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['MediumOffice - ClosedOffice'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['MediumOffice - Conference'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['MediumOffice - Corridor'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['MediumOffice - Elec/MechRoom'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['MediumOffice - IT_Room'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['MediumOffice - Lobby'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['MediumOffice - OpenOffice'] = { ratio: 0.99, space_type_gen: true, default: true }
+          hash['MediumOffice - Restroom'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['MediumOffice - Stair'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['MediumOffice - Storage'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['MediumOffice - Classroom'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['MediumOffice - Dining'] = { ratio: 0.99, space_type_gen: true, default: false }
+          hash['WholeBuilding - Md Office'] = { ratio: 0.0, space_type_gen: true, default: false }
+        end
+      elsif building_type == 'LargeOffice'
+        if ['DOE Ref Pre-1980', 'DOE Ref 1980-2004'].include?(template)
+          if whole_building
+            hash['WholeBuilding - Lg Office'] = { ratio: 1.0, space_type_gen: true, default: true }
+          else
+            hash['BreakRoom'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['ClosedOffice'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Conference'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Corridor'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Elec/MechRoom'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['IT_Room'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Lobby'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['OpenOffice'] = { ratio: 0.99, space_type_gen: true, default: true }
+            hash['PrintRoom'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Restroom'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Stair'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Storage'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Vending'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['WholeBuilding - Lg Office'] = { ratio: 0.0, space_type_gen: true, default: false }
+          end
+        else
+          if whole_building
+            hash['WholeBuilding - Lg Office'] = { ratio: 0.9737, space_type_gen: true, default: true }
+            hash['OfficeLarge Data Center'] = { ratio: 0.0094, space_type_gen: true, default: false }
+            hash['OfficeLarge Main Data Center'] = { ratio: 0.0169, space_type_gen: true, default: false }
+          else
+            hash['BreakRoom'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['ClosedOffice'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Conference'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Corridor'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Elec/MechRoom'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['IT_Room'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Lobby'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['OpenOffice'] = { ratio: 0.99, space_type_gen: true, default: true }
+            hash['PrintRoom'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Restroom'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Stair'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Storage'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['Vending'] = { ratio: 0.99, space_type_gen: true, default: false }
+            hash['WholeBuilding - Lg Office'] = { ratio: 0.0, space_type_gen: true, default: false }
+            hash['OfficeLarge Data Center'] = { ratio: 0.0, space_type_gen: true, default: false }
+            hash['OfficeLarge Main Data Center'] = { ratio: 0.0, space_type_gen: true, default: false }
+          end
+        end
+      elsif building_type == 'SmallHotel'
+        if ['DOE Ref Pre-1980', 'DOE Ref 1980-2004'].include?(template)
+          hash['Corridor'] = { ratio: 0.1313, space_type_gen: true, default: false }
+          hash['Elec/MechRoom'] = { ratio: 0.0038, space_type_gen: true, default: false }
+          hash['ElevatorCore'] = { ratio: 0.0113, space_type_gen: true, default: false }
+          hash['Exercise'] = { ratio: 0.0081, space_type_gen: true, default: false }
+          hash['GuestLounge'] = { ratio: 0.0406, space_type_gen: true, default: false }
+          hash['GuestRoom'] = { ratio: 0.6313, space_type_gen: true, default: true }
+          hash['Laundry'] = { ratio: 0.0244, space_type_gen: true, default: false }
+          hash['Mechanical'] = { ratio: 0.0081, space_type_gen: true, default: false }
+          hash['Meeting'] = { ratio: 0.0200, space_type_gen: true, default: false }
+          hash['Office'] = { ratio: 0.0325, space_type_gen: true, default: false }
+          hash['PublicRestroom'] = { ratio: 0.0081, space_type_gen: true, default: false }
+          hash['StaffLounge'] = { ratio: 0.0081, space_type_gen: true, default: false }
+          hash['Stair'] = { ratio: 0.0400, space_type_gen: true, default: false }
+          hash['Storage'] = { ratio: 0.0325, space_type_gen: true, default: false }
+        else
+          hash['Corridor'] = { ratio: 0.1313, space_type_gen: true, default: false }
+          hash['Elec/MechRoom'] = { ratio: 0.0038, space_type_gen: true, default: false }
+          hash['ElevatorCore'] = { ratio: 0.0113, space_type_gen: true, default: false }
+          hash['Exercise'] = { ratio: 0.0081, space_type_gen: true, default: false }
+          hash['GuestLounge'] = { ratio: 0.0406, space_type_gen: true, default: false }
+          hash['GuestRoom123Occ'] = { ratio: 0.4081, space_type_gen: true, default: true }
+          hash['GuestRoom123Vac'] = { ratio: 0.2231, space_type_gen: true, default: false }
+          hash['Laundry'] = { ratio: 0.0244, space_type_gen: true, default: false }
+          hash['Mechanical'] = { ratio: 0.0081, space_type_gen: true, default: false }
+          hash['Meeting'] = { ratio: 0.0200, space_type_gen: true, default: false }
+          hash['Office'] = { ratio: 0.0325, space_type_gen: true, default: false }
+          hash['PublicRestroom'] = { ratio: 0.0081, space_type_gen: true, default: false }
+          hash['StaffLounge'] = { ratio: 0.0081, space_type_gen: true, default: false }
+          hash['Stair'] = { ratio: 0.0400, space_type_gen: true, default: false }
+          hash['Storage'] = { ratio: 0.0325, space_type_gen: true, default: false }
+        end
+      elsif building_type == 'LargeHotel'
+        hash['Banquet'] = { ratio: 0.0585, space_type_gen: true, default: false }
+        hash['Basement'] = { ratio: 0.1744, space_type_gen: false, default: false }
+        hash['Cafe'] = { ratio: 0.0166, space_type_gen: true, default: false }
+        hash['Corridor'] = { ratio: 0.1736, space_type_gen: true, default: false }
+        hash['GuestRoom'] = { ratio: 0.4099, space_type_gen: true, default: true }
+        hash['Kitchen'] = { ratio: 0.0091, space_type_gen: true, default: false }
+        hash['Laundry'] = { ratio: 0.0069, space_type_gen: true, default: false }
+        hash['Lobby'] = { ratio: 0.1153, space_type_gen: true, default: false }
+        hash['Mechanical'] = { ratio: 0.0145, space_type_gen: true, default: false }
+        hash['Retail'] = { ratio: 0.0128, space_type_gen: true, default: false }
+        hash['Storage'] = { ratio: 0.0084, space_type_gen: true, default: false }
+      elsif building_type == 'Warehouse'
+        hash['Bulk'] = { ratio: 0.6628, space_type_gen: true, default: true }
+        hash['Fine'] = { ratio: 0.2882, space_type_gen: true, default: false }
+        hash['Office'] = { ratio: 0.0490, space_type_gen: true, default: false }
+      elsif building_type == 'RetailStandalone'
+        hash['Back_Space'] = { ratio: 0.1656, space_type_gen: true, default: false }
+        hash['Entry'] = { ratio: 0.0052, space_type_gen: true, default: false }
+        hash['Point_of_Sale'] = { ratio: 0.0657, space_type_gen: true, default: false }
+        hash['Retail'] = { ratio: 0.7635, space_type_gen: true, default: true }
+      elsif building_type == 'RetailStripmall'
+        hash['Strip mall - type 1'] = { ratio: 0.25, space_type_gen: true, default: false }
+        hash['Strip mall - type 2'] = { ratio: 0.25, space_type_gen: true, default: false }
+        hash['Strip mall - type 3'] = { ratio: 0.50, space_type_gen: true, default: true }
+      elsif building_type == 'QuickServiceRestaurant'
+        hash['Dining'] = { ratio: 0.5, space_type_gen: true, default: true }
+        hash['Kitchen'] = { ratio: 0.5, space_type_gen: true, default: false }
+      elsif building_type == 'FullServiceRestaurant'
+        hash['Dining'] = { ratio: 0.7272, space_type_gen: true, default: true }
+        hash['Kitchen'] = { ratio: 0.2728, space_type_gen: true, default: false }
+      elsif building_type == 'MidriseApartment'
+        hash['Apartment'] = { ratio: 0.8727, space_type_gen: true, default: true }
+        hash['Corridor'] = { ratio: 0.0991, space_type_gen: true, default: false }
+        hash['Office'] = { ratio: 0.0282, space_type_gen: true, default: false }
+      elsif building_type == 'HighriseApartment'
+        hash['Apartment'] = { ratio: 0.8896, space_type_gen: true, default: true }
+        hash['Corridor'] = { ratio: 0.0991, space_type_gen: true, default: false }
+        hash['Office'] = { ratio: 0.0113, space_type_gen: true, default: false }
+      elsif building_type == 'Hospital'
+        hash['Basement'] = { ratio: 0.1667, space_type_gen: false, default: false }
+        hash['Corridor'] = { ratio: 0.1741, space_type_gen: true, default: false }
+        hash['Dining'] = { ratio: 0.0311, space_type_gen: true, default: false }
+        hash['ER_Exam'] = { ratio: 0.0099, space_type_gen: true, default: false }
+        hash['ER_NurseStn'] = { ratio: 0.0551, space_type_gen: true, default: false }
+        hash['ER_Trauma'] = { ratio: 0.0025, space_type_gen: true, default: false }
+        hash['ER_Triage'] = { ratio: 0.0050, space_type_gen: true, default: false }
+        hash['ICU_NurseStn'] = { ratio: 0.0298, space_type_gen: true, default: false }
+        hash['ICE_Open'] = { ratio: 0.0275, space_type_gen: true, default: false }
+        hash['ICU_PatRm'] = { ratio: 0.0115, space_type_gen: true, default: false }
+        hash['Kitchen'] = { ratio: 0.0414, space_type_gen: true, default: false }
+        hash['Lab'] = { ratio: 0.0236, space_type_gen: true, default: false }
+        hash['Lobby'] = { ratio: 0.0657, space_type_gen: true, default: false }
+        hash['NurseStn'] = { ratio: 0.1723, space_type_gen: true, default: false }
+        hash['Office'] = { ratio: 0.0286, space_type_gen: true, default: false }
+        hash['OR'] = { ratio: 0.0273, space_type_gen: true, default: false }
+        hash['PatCorridor'] = { ratio: 0.0, space_type_gen: true, default: false } # not in prototype
+        hash['PatRoom'] = { ratio: 0.0845, space_type_gen: true, default: true }
+        hash['PhysTherapy'] = { ratio: 0.0217, space_type_gen: true, default: false }
+        hash['Radiology'] = { ratio: 0.0217, space_type_gen: true, default: false }
+      elsif building_type == 'Outpatient'
+        hash['Anesthesia'] = { ratio: 0.0026, space_type_gen: true, default: false }
+        hash['BioHazard'] = { ratio: 0.0014, space_type_gen: true, default: false }
+        hash['Cafe'] = { ratio: 0.0103, space_type_gen: true, default: false }
+        hash['CleanWork'] = { ratio: 0.0071, space_type_gen: true, default: false }
+        hash['Conference'] = { ratio: 0.0082, space_type_gen: true, default: false }
+        hash['DresingRoom'] = { ratio: 0.0021, space_type_gen: true, default: false }
+        hash['Elec/MechRoom'] = { ratio: 0.0109, space_type_gen: true, default: false }
+        hash['ElevatorPumpRoom'] = { ratio: 0.0022, space_type_gen: true, default: false }
+        hash['Exam'] = { ratio: 0.1029, space_type_gen: true, default: true }
+        hash['Hall'] = { ratio: 0.1924, space_type_gen: true, default: false }
+        hash['IT_Room'] = { ratio: 0.0027, space_type_gen: true, default: false }
+        hash['Janitor'] = { ratio: 0.0672, space_type_gen: true, default: false }
+        hash['Lobby'] = { ratio: 0.0152, space_type_gen: true, default: false }
+        hash['LockerRoom'] = { ratio: 0.0190, space_type_gen: true, default: false }
+        hash['Lounge'] = { ratio: 0.0293, space_type_gen: true, default: false }
+        hash['MedGas'] = { ratio: 0.0014, space_type_gen: true, default: false }
+        hash['MRI'] = { ratio: 0.0107, space_type_gen: true, default: false }
+        hash['MRI_Control'] = { ratio: 0.0041, space_type_gen: true, default: false }
+        hash['NurseStation'] = { ratio: 0.0189, space_type_gen: true, default: false }
+        hash['Office'] = { ratio: 0.1828, space_type_gen: true, default: false }
+        hash['OR'] = { ratio: 0.0346, space_type_gen: true, default: false }
+        hash['PACU'] = { ratio: 0.0232, space_type_gen: true, default: false }
+        hash['PhysicalTherapy'] = { ratio: 0.0462, space_type_gen: true, default: false }
+        hash['PreOp'] = { ratio: 0.0129, space_type_gen: true, default: false }
+        hash['ProcedureRoom'] = { ratio: 0.0070, space_type_gen: true, default: false }
+        hash['Reception'] = { ratio: 0.0365, space_type_gen: true, default: false }
+        hash['Soil Work'] = { ratio: 0.0088, space_type_gen: true, default: false }
+        hash['Stair'] = { ratio: 0.0146, space_type_gen: true, default: false }
+        hash['Toilet'] = { ratio: 0.0193, space_type_gen: true, default: false }
+        hash['Undeveloped'] = { ratio: 0.0835, space_type_gen: false, default: false }
+        hash['Xray'] = { ratio: 0.0220, space_type_gen: true, default: false }
+      elsif building_type == 'SuperMarket'
+        # TODO: - populate ratios for SuperMarket
+        hash['Bakery'] = { ratio: 0.99, space_type_gen: true, default: false }
+        hash['Deli'] = { ratio: 0.99, space_type_gen: true, default: false }
+        hash['DryStorage'] = { ratio: 0.99, space_type_gen: true, default: false }
+        hash['Office'] = { ratio: 0.99, space_type_gen: true, default: false }
+        hash['Produce'] = { ratio: 0.99, space_type_gen: true, default: true }
+        hash['Sales'] = { ratio: 0.99, space_type_gen: true, default: true }
+        hash['Corridor'] = { ratio: 0.99, space_type_gen: true, default: true }
+        hash['Dining'] = { ratio: 0.99, space_type_gen: true, default: true }
+        hash['Elec/MechRoom'] = { ratio: 0.99, space_type_gen: true, default: true }
+        hash['Meeting'] = { ratio: 0.99, space_type_gen: true, default: true }
+        hash['Restroom'] = { ratio: 0.99, space_type_gen: true, default: true }
+        hash['Vestibule'] = { ratio: 0.99, space_type_gen: true, default: true }
+      else
+        return false
+      end
+
+      return hash
+
+    end
+
 
     def set_bldg_and_system_type(occupancy_type, total_floor_area, raise_exception)
       # DOE Prototype building types:from openstudio-standards/lib/openstudio-standards/prototypes/common/prototype_metaprogramming.rb
